+++ firebase-messaging/CHANGELOG.md
@@ -1,9 +1,10 @@
 # Unreleased
 
+- [feature] Added support for FCM registration using Firebase Installation ID as token
+
 # 25.0.2
 
 - [changed] Fix ANR in SharedPreferencesQueue by reducing lock contention
-- [feature] Added support for FCM registration using Firebase Installation ID as token
 - [deprecated] Existing getToken, deleteToken, onNewToken api(s) are deprecated
 
 # 25.0.1